server/spike/adk/requirements.txt
google-adk>=1.31.0
google-genai>=1.0.0
